weather-app-source/services/api-gateway/requirements.txt
Flask==2.3.2
requests==2.31.0
protobuf>=4.21.0
grpcio>=1.50.0
grpcio-tools>=1.50.0